refactor: rename folders to avoid conflicts
