+++ simulations/vip-463/bsctestnet.ts
@@ -1,11 +1,11 @@
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip460 from "../../vips/vip-460/bsctestnet";
+import vip463 from "../../vips/vip-463/bsctestnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(47551000, async () => {
-  testVip("VIP-460", await vip460(), {
+  testVip("VIP-463", await vip463(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
