+++ simulations/vip-463/unichainsepolia.ts
@@ -8,7 +8,7 @@ import { checkIsolatedPoolsComptrollers } from "src/vip-framework/checks/checkIs
 import { checkVToken } from "src/vip-framework/checks/checkVToken";
 import { checkInterestRate } from "src/vip-framework/checks/interestRateModel";
 
-import vip460, { COMPTROLLER_CORE, UNI, VUNI_CORE } from "../../vips/vip-460/bsctestnet-uni";
+import vip463, { COMPTROLLER_CORE, UNI, VUNI_CORE } from "../../vips/vip-463/bsctestnet-uni";
 import COMPTROLLER_ABI from "./abi/comptroller.json";
 import VTOKEN_ABI from "./abi/vToken.json";
 
@@ -33,7 +33,7 @@ forking(13205017, async () => {
     });
   });
 
-  testForkedNetworkVipCommands("add UNI market", await vip460());
+  testForkedNetworkVipCommands("add UNI market", await vip463());
   describe("Post-Execution state", () => {
     let interestRateModelAddresses: string;
     let vToken: Contract;
